# Speed up creating Mark objects in the Cython version