@bdraco 2025-05-16
Still a nice speed up even with the helper. Note that 3.1 got a bit faster already since I opened this PR

### With Cython 3.1 `main`

                           Benchmark Results                             
┏━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━┓
┃             Benchmark ┃   Time (best) ┃ Rel. StdDev ┃ Run time ┃ Iters ┃
┡━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━┩
│ test_large_parse_yaml │ 177,877,584ns │        0.7% │    2.88s │    16 │
└───────────────────────┴───────────────┴─────────────┴──────────┴───────┘
  

### With Cython 3.1 `64e40d89ddb54b192d84132e6540fcc6553a8050` (original commit)

                          Benchmark Results                             
┏━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━┓
┃             Benchmark ┃   Time (best) ┃ Rel. StdDev ┃ Run time ┃ Iters ┃
┡━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━┩
│ test_large_parse_yaml │ 159,839,750ns │        1.4% │    2.94s │    18 │
└───────────────────────┴───────────────┴─────────────┴──────────┴───────┘

### With Cython 3.1 (added _create_mark) `964c641aad18b5df3ef92b9202a7b82283939d9e` (original commit)

                            Benchmark Results                             
┏━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━┓
┃             Benchmark ┃   Time (best) ┃ Rel. StdDev ┃ Run time ┃ Iters ┃
┡━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━┩
│ test_large_parse_yaml │ 162,833,125ns │        0.8% │    2.97s │    18 │
└───────────────────────┴───────────────┴─────────────┴──────────┴───────┘