@bdraco 2025-05-16
Pushed another one that a bit different. Single call point but it does have a bit more maintenance burden. 688ebe766c4c9e4cb4e3d79aa9cfc5fbb9c0f097

```
                            Benchmark Results                             
┏━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━┓
┃             Benchmark ┃   Time (best) ┃ Rel. StdDev ┃ Run time ┃ Iters ┃
┡━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━┩
│ test_large_parse_yaml │ 155,708,875ns │        0.7% │    2.84s │    18 │
└───────────────────────┴───────────────┴─────────────┴──────────┴───────┘
```